P3X-294: Extracts method to convert file on service layer
