--- .travis.yml
@@ -1,1 +1,0 @@
-language: node_js
