+++ routes/main.js
@@ -62,105 +62,19 @@ module.exports = function (app) {
             }
             var filePath = files.fileUpload.filepath;
             var fileType = path.extname(files.fileUpload.originalFilename);
-            console.log("File Path: " + filePath);
-            console.log("File Type: " + fileType);
 
             const sep = fields.sep;
             const fixation = fields.fixation;
 
-            // Save the file on the server code.
+            var fileData = fs.readFileSync(filePath);
 
-            // var newPath = path.join(__dirname + '/../public/files') + '/'+files.fileUpload.originalFilename;
-            // var test = fs.readFileSync(oldPath);
-            // fs.writeFile(newPath, test, function(err){
-
-            var fileData = fs.readFileSync(filePath, function(err, data){
-                if(err){
-                    console.error(err);
-                    return;
-                }
-            });
-
-            // TXT files
-            if(fileType == '.txt'){
-                var text = fileData.toString();
-
-                console.log("Text to be converted with text-vide: ", text, " with a fixation of: ", fixation);
-                const textWithBionic = bionicReaderService.convertTextUsingTextVide(text, sep, fixation) 
-                res.render('customise.ejs', {text, textWithBionic, fixation});                 
-            }
-
-            // PDF files
-            else if(fileType == '.pdf'){
-                const pdf = require('pdf-parse');
-
-                pdf(fileData).then(function(data){
-                    var text = data.text;
-        
-                    console.log("Text to be converted with text-vide: ", text, " with a fixation of: ", fixation);     
-                    const textWithBionic = bionicReaderService.convertTextUsingTextVide(text, sep, fixation)
+            bionicReaderService.convertFile(fileData, fileType, {sep, fixation})
+                .then(({text, textWithBionic, fixation}) => {
                     res.render('customise.ejs', {text, textWithBionic, fixation});
-                });                   
-            }
-
-            // DOCX files
-            else if(fileType == '.docx'){
-                const mammoth = require('mammoth');
-                
-                mammoth.extractRawText({buffer: fileData}).then(function(result){
-                    var text = result.value;
-
-                    console.log("Text to be converted with text-vide: ", text, " with a fixation of: ", fixation);
-                    const textWithBionic = bionicReaderService.convertTextUsingTextVide(text, sep, fixation) 
-                    res.render('customise.ejs', {text, textWithBionic, fixation});                    
-                });
-
-            }
-
-            // RTF files
-            else if(fileType == '.rtf'){
-                const rtfParser = require('rtf-parser');
-
-                rtfParser.stream(fs.createReadStream(filePath), function(err, doc){
-                    if(err){
-                        console.error(err);
-                        return;
-                    }
-
-                    var text = "";
-                    for(var i = 0; i < doc.content.length; i++){
-                        for(var j = 0; j < doc.content[i].content.length; j++){
-                            text += doc.content[i].content[j].value;
-                        }
-                        text+= '\n';
-                    }
-
-                    console.log("Text to be converted with text-vide: ", text, " with a fixation of: ", fixation);
-                    const textWithBionic = bionicReaderService.convertTextUsingTextVide(text, sep, fixation) 
-                    res.render('customise.ejs', {text, textWithBionic, fixation}); 
-                });
-            }
-
-            // HTML files
-            else if(fileType == '.html'){
-                const { convert } = require('html-to-text');
-
-                var text = convert(fileData)
-                console.log(text);
-
-                console.log("Text to be converted with text-vide: ", text, " with a fixation of: ", fixation);
-                const textWithBionic = bionicReaderService.convertTextUsingTextVide(text, sep, fixation) 
-                res.render('customise.ejs', {text, textWithBionic, fixation}); 
-            }
-
-            else{
-                console.error(fileType + " is not supported!");
-            }
-            
-            // console.log("Text to be converted with text-vide: ", text, " with a fixation of: ", fixation);
-            // const textWithBionic = bionicReaderService.convertTextUsingTextVide(text, sep, fixation)
-            // res.render('customise.ejs', {text, textWithBionic, fixation});
-     
+                })
+                .catch((err) => {
+                    console.log(err);
+                })
         });
     });
